+++ Assets/Project/Scripts/Enemies/EnemyModel.cs
@@ -8,11 +8,11 @@ namespace Project.Scripts.Enemies
 {
     public class EnemyModel
     {
-        public float EXP { get; private set; }
+        public int EXP { get; private set; }
         private Weapon<StoneCannonConfig> CurrentWeapon { get; set; }
         public Health EnemyHealth { get; private set; }
 
-        public EnemyModel(EnemyConfig config, Weapon<StoneCannonConfig> weapon, Health health, float eXP)
+        public EnemyModel(EnemyConfig config, Weapon<StoneCannonConfig> weapon, Health health, int eXP)
         {
             CurrentWeapon = weapon;
             EnemyHealth = health;
